+++ checkpoint/orbax/checkpoint/_src/testing/test_tree_utils.py
@@ -20,8 +20,20 @@ import chex
 import jax
 import jax.numpy as jnp
 import numpy as np
+from orbax.checkpoint._src.metadata import tree as tree_metadata
+from orbax.checkpoint._src.metadata import tree_rich_types
 
 PyTree = Any
+
+
+def _build_namedtuple(
+    cls,
+    field_value_tuples: list[tuple[str, tree_metadata.ValueMetadataEntry]],
+):
+  fields, values = zip(*field_value_tuples)
+  module_name, class_name = tree_rich_types._module_and_class_name(cls)  # pylint: disable=protected-access
+  new_type = tree_rich_types._new_namedtuple_type(module_name, class_name, fields)  # pylint: disable=protected-access
+  return new_type(*values)
 
 
 class MuNu(NamedTuple):
@@ -125,6 +137,9 @@ class TestPyTree:
       BaseException, a function to provide the pytree or None. None implies that
       expected restored tree should be the same as the tree returned by
       `provide_tree` function.
+    expected_nested_tree_metadata_with_rich_types: PyTree of ValueMetadataEntry
+      as returned by `InternalTreeMetadata.as_nested_tree` with
+      `PyTreeMetadataOptions.support_rich_types=true`.
   """
 
   unique_name: str
@@ -134,6 +149,7 @@ class TestPyTree:
   expected_tree_metadata_key_json: Dict[str, Any]
   expected_save_response: BaseException | None = None
   expected_restore_response: BaseException | Callable[[], PyTree] | None = None
+  expected_nested_tree_metadata_with_rich_types: PyTree | None = None
 
   def __post_init__(self):
     self.expected_restore_response = (
@@ -162,6 +178,12 @@ TEST_PYTREES = [
         },
         expected_save_response=ValueError(),
         expected_restore_response=ValueError(),
+        expected_nested_tree_metadata_with_rich_types=(
+            tree_metadata.ValueMetadataEntry(
+                value_type='Dict',
+                skip_deserialize=True,
+            )
+        ),
     ),
     TestPyTree(
         unique_name='empty_dict',
@@ -174,6 +196,12 @@ TEST_PYTREES = [
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'empty_dict': tree_metadata.ValueMetadataEntry(
+                value_type='Dict',
+                skip_deserialize=True,
+            )
         },
     ),
     TestPyTree(
@@ -188,6 +216,12 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'empty_list': tree_metadata.ValueMetadataEntry(
+                value_type='List',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='empty_tuple',
@@ -201,6 +235,12 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'empty_tuple': tree_metadata.ValueMetadataEntry(
+                value_type='Tuple',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='empty_named_tuple',
@@ -209,11 +249,16 @@ TEST_PYTREES = [
             "('empty_named_tuple',)": {
                 'key_metadata': ({'key': 'empty_named_tuple', 'key_type': 2},),
                 'value_metadata': {
-                    # TODO: b/365169723 - Handle empty NamedTuple.
                     'value_type': 'None',
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'empty_named_tuple': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
         },
     ),
     TestPyTree(
@@ -231,6 +276,14 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'tuple_of_empty_list': tuple([
+                tree_metadata.ValueMetadataEntry(
+                    value_type='List',
+                    skip_deserialize=True,
+                ),
+            ])
+        },
     ),
     TestPyTree(
         unique_name='list_of_empty_tuple',
@@ -247,6 +300,14 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'list_of_empty_tuple': [
+                tree_metadata.ValueMetadataEntry(
+                    value_type='Tuple',
+                    skip_deserialize=True,
+                ),
+            ]
+        },
     ),
     TestPyTree(
         unique_name='none_param',
@@ -260,6 +321,12 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'none_param': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='scalar_param',
@@ -272,6 +339,12 @@ TEST_PYTREES = [
                     'skip_deserialize': False,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'scalar_param': tree_metadata.ValueMetadataEntry(
+                value_type='scalar',
+                skip_deserialize=False,
+            )
         },
     ),
     TestPyTree(
@@ -301,6 +374,18 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'b': {
+                'scalar_param': tree_metadata.ValueMetadataEntry(
+                    value_type='scalar',
+                    skip_deserialize=False,
+                ),
+                'nested_scalar_param': tree_metadata.ValueMetadataEntry(
+                    value_type='scalar',
+                    skip_deserialize=False,
+                ),
+            }
+        },
     ),
     TestPyTree(
         unique_name='list_of_arrays',
@@ -326,6 +411,152 @@ TEST_PYTREES = [
                     'skip_deserialize': False,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'list_of_arrays': [
+                tree_metadata.ValueMetadataEntry(
+                    value_type='np.ndarray',
+                    skip_deserialize=False,
+                ),
+                tree_metadata.ValueMetadataEntry(
+                    value_type='jax.Array',
+                    skip_deserialize=False,
+                ),
+            ]
+        },
+    ),
+    TestPyTree(
+        unique_name='dict_of_nested_data',
+        provide_tree=lambda: {
+            'x': {'a': np.arange(8), 'b': MyEmptyChex()},
+            'y': (np.arange(8), MyChex(my_np_array=np.arange(8))),
+            'z': [
+                {'c': np.arange(8)},
+                [
+                    (np.arange(8),),
+                ],
+            ],
+        },
+        expected_tree_metadata_key_json={
+            "('x', 'a')": {
+                'key_metadata': (
+                    {'key': 'x', 'key_type': 2},
+                    {'key': 'a', 'key_type': 2},
+                ),
+                'value_metadata': {
+                    'value_type': 'np.ndarray',
+                    'skip_deserialize': False,
+                },
+            },
+            "('x', 'b')": {
+                'key_metadata': (
+                    {'key': 'x', 'key_type': 2},
+                    {'key': 'b', 'key_type': 2},
+                ),
+                'value_metadata': {
+                    'value_type': 'Dict',
+                    'skip_deserialize': True,
+                },
+            },
+            "('y', '0')": {
+                'key_metadata': (
+                    {'key': 'y', 'key_type': 2},
+                    {'key': '0', 'key_type': 1},
+                ),
+                'value_metadata': {
+                    'value_type': 'np.ndarray',
+                    'skip_deserialize': False,
+                },
+            },
+            "('y', '1', 'my_jax_array')": {
+                'key_metadata': (
+                    {'key': 'y', 'key_type': 2},
+                    {'key': '1', 'key_type': 1},
+                    {'key': 'my_jax_array', 'key_type': 2},
+                ),
+                'value_metadata': {
+                    'value_type': 'None',
+                    'skip_deserialize': True,
+                },
+            },
+            "('y', '1', 'my_np_array')": {
+                'key_metadata': (
+                    {'key': 'y', 'key_type': 2},
+                    {'key': '1', 'key_type': 1},
+                    {'key': 'my_np_array', 'key_type': 2},
+                ),
+                'value_metadata': {
+                    'value_type': 'np.ndarray',
+                    'skip_deserialize': False,
+                },
+            },
+            "('z', '0', 'c')": {
+                'key_metadata': (
+                    {'key': 'z', 'key_type': 2},
+                    {'key': '0', 'key_type': 1},
+                    {'key': 'c', 'key_type': 2},
+                ),
+                'value_metadata': {
+                    'value_type': 'np.ndarray',
+                    'skip_deserialize': False,
+                },
+            },
+            "('z', '1', '0', '0')": {
+                'key_metadata': (
+                    {'key': 'z', 'key_type': 2},
+                    {'key': '1', 'key_type': 1},
+                    {'key': '0', 'key_type': 1},
+                    {'key': '0', 'key_type': 1},
+                ),
+                'value_metadata': {
+                    'value_type': 'np.ndarray',
+                    'skip_deserialize': False,
+                },
+            },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'x': {
+                'a': tree_metadata.ValueMetadataEntry(
+                    value_type='np.ndarray',
+                    skip_deserialize=False,
+                ),
+                'b': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+            },
+            'y': (
+                tree_metadata.ValueMetadataEntry(
+                    value_type='np.ndarray',
+                    skip_deserialize=False,
+                ),
+                {
+                    'my_jax_array': tree_metadata.ValueMetadataEntry(
+                        value_type='None',
+                        skip_deserialize=True,
+                    ),
+                    'my_np_array': tree_metadata.ValueMetadataEntry(
+                        value_type='np.ndarray',
+                        skip_deserialize=False,
+                    ),
+                },
+            ),
+            'z': [
+                {
+                    'c': tree_metadata.ValueMetadataEntry(
+                        value_type='np.ndarray',
+                        skip_deserialize=False,
+                    )
+                },
+                [
+                    (
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='np.ndarray',
+                            skip_deserialize=False,
+                        ),
+                    ),
+                ],
+            ],
         },
     ),
     TestPyTree(
@@ -353,6 +584,18 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'tuple_of_arrays': tuple([
+                tree_metadata.ValueMetadataEntry(
+                    value_type='np.ndarray',
+                    skip_deserialize=False,
+                ),
+                tree_metadata.ValueMetadataEntry(
+                    value_type='jax.Array',
+                    skip_deserialize=False,
+                ),
+            ])
+        },
     ),
     TestPyTree(
         unique_name='mu_nu',
@@ -379,6 +622,27 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'mu_nu': _build_namedtuple(
+                MuNu,
+                [
+                    (
+                        'mu',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='jax.Array',
+                            skip_deserialize=False,
+                        ),
+                    ),
+                    (
+                        'nu',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='np.ndarray',
+                            skip_deserialize=False,
+                        ),
+                    ),
+                ],
+            )
+        },
     ),
     TestPyTree(
         unique_name='my_empty_chex',
@@ -391,6 +655,12 @@ TEST_PYTREES = [
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'my_empty_chex': tree_metadata.ValueMetadataEntry(
+                value_type='Dict',
+                skip_deserialize=True,
+            )
         },
     ),
     TestPyTree(
@@ -417,6 +687,18 @@ TEST_PYTREES = [
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'default_my_chex': {
+                'my_jax_array': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+                'my_np_array': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+            }
         },
     ),
     TestPyTree(
@@ -448,6 +730,18 @@ TEST_PYTREES = [
                     'skip_deserialize': False,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'my_chex': {
+                'my_jax_array': tree_metadata.ValueMetadataEntry(
+                    value_type='jax.Array',
+                    skip_deserialize=False,
+                ),
+                'my_np_array': tree_metadata.ValueMetadataEntry(
+                    value_type='np.ndarray',
+                    skip_deserialize=False,
+                ),
+            }
         },
     ),
     TestPyTree(
@@ -496,6 +790,26 @@ TEST_PYTREES = [
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'my_chex_with_nested_attrs': {
+                'my_chex': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+                'my_dict': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+                'my_empty_chex': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+                'my_list': tree_metadata.ValueMetadataEntry(
+                    value_type='None',
+                    skip_deserialize=True,
+                ),
+            }
         },
     ),
     TestPyTree(
@@ -588,6 +902,44 @@ TEST_PYTREES = [
                 },
             },
         },
+        expected_nested_tree_metadata_with_rich_types={
+            'my_chex_with_nested_attrs': {
+                'my_chex': {
+                    'my_jax_array': tree_metadata.ValueMetadataEntry(
+                        value_type='jax.Array',
+                        skip_deserialize=False,
+                    ),
+                    'my_np_array': tree_metadata.ValueMetadataEntry(
+                        value_type='np.ndarray',
+                        skip_deserialize=False,
+                    ),
+                },
+                'my_dict': {
+                    'a': tree_metadata.ValueMetadataEntry(
+                        value_type='jax.Array',
+                        skip_deserialize=False,
+                    ),
+                    'b': tree_metadata.ValueMetadataEntry(
+                        value_type='np.ndarray',
+                        skip_deserialize=False,
+                    ),
+                },
+                'my_empty_chex': tree_metadata.ValueMetadataEntry(
+                    value_type='Dict',
+                    skip_deserialize=True,
+                ),
+                'my_list': [
+                    tree_metadata.ValueMetadataEntry(
+                        value_type='jax.Array',
+                        skip_deserialize=False,
+                    ),
+                    tree_metadata.ValueMetadataEntry(
+                        value_type='np.ndarray',
+                        skip_deserialize=False,
+                    ),
+                ],
+            }
+        },
     ),
     TestPyTree(
         unique_name='my_empty_class',
@@ -603,6 +955,12 @@ TEST_PYTREES = [
         },
         expected_save_response=ValueError(),
         expected_restore_response=ValueError(),
+        expected_nested_tree_metadata_with_rich_types={
+            'my_empty_class': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='default_my_class',
@@ -618,6 +976,12 @@ TEST_PYTREES = [
         },
         expected_save_response=ValueError(),
         expected_restore_response=ValueError(),
+        expected_nested_tree_metadata_with_rich_types={
+            'my_class': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='my_class',
@@ -635,6 +999,12 @@ TEST_PYTREES = [
         },
         expected_save_response=ValueError(),
         expected_restore_response=ValueError(),
+        expected_nested_tree_metadata_with_rich_types={
+            'my_class': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='default_my_class_with_nested_attrs',
@@ -654,6 +1024,12 @@ TEST_PYTREES = [
         },
         expected_save_response=ValueError(),
         expected_restore_response=ValueError(),
+        expected_nested_tree_metadata_with_rich_types={
+            'my_class_with_nested_attrs': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='my_class_with_nested_attrs',
@@ -680,6 +1056,12 @@ TEST_PYTREES = [
         },
         expected_save_response=ValueError(),
         expected_restore_response=ValueError(),
+        expected_nested_tree_metadata_with_rich_types={
+            'my_class_with_nested_attrs': tree_metadata.ValueMetadataEntry(
+                value_type='None',
+                skip_deserialize=True,
+            )
+        },
     ),
     TestPyTree(
         unique_name='default_named_tuple_with_nested_attrs',
@@ -727,6 +1109,41 @@ TEST_PYTREES = [
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'named_tuple_with_nested_attrs': _build_namedtuple(
+                NamedTupleWithNestedAttributes,
+                [
+                    (
+                        'nested_mu_nu',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='None',
+                            skip_deserialize=True,
+                        ),
+                    ),
+                    (
+                        'nested_dict',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='None',
+                            skip_deserialize=True,
+                        ),
+                    ),
+                    (
+                        'nested_tuple',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='None',
+                            skip_deserialize=True,
+                        ),
+                    ),
+                    (
+                        'nested_empty_named_tuple',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='None',
+                            skip_deserialize=True,
+                        ),
+                    ),
+                ],
+            )
         },
     ),
     TestPyTree(
@@ -812,11 +1229,72 @@ TEST_PYTREES = [
                     {'key': 'nested_empty_named_tuple', 'key_type': 2},
                 ),
                 'value_metadata': {
-                    # TODO: b/365169723 - Handle empty NamedTuple.
                     'value_type': 'None',
                     'skip_deserialize': True,
                 },
             },
+        },
+        expected_nested_tree_metadata_with_rich_types={
+            'named_tuple_with_nested_attrs': _build_namedtuple(
+                NamedTupleWithNestedAttributes,
+                [
+                    (
+                        'nested_mu_nu',
+                        _build_namedtuple(
+                            MuNu,
+                            [
+                                (
+                                    'mu',
+                                    tree_metadata.ValueMetadataEntry(
+                                        value_type='jax.Array',
+                                        skip_deserialize=False,
+                                    ),
+                                ),
+                                (
+                                    'nu',
+                                    tree_metadata.ValueMetadataEntry(
+                                        value_type='np.ndarray',
+                                        skip_deserialize=False,
+                                    ),
+                                ),
+                            ],
+                        ),
+                    ),
+                    (
+                        'nested_dict',
+                        {
+                            'a': tree_metadata.ValueMetadataEntry(
+                                value_type='jax.Array',
+                                skip_deserialize=False,
+                            ),
+                            'b': tree_metadata.ValueMetadataEntry(
+                                value_type='np.ndarray',
+                                skip_deserialize=False,
+                            ),
+                        },
+                    ),
+                    (
+                        'nested_tuple',
+                        (
+                            tree_metadata.ValueMetadataEntry(
+                                value_type='jax.Array',
+                                skip_deserialize=False,
+                            ),
+                            tree_metadata.ValueMetadataEntry(
+                                value_type='np.ndarray',
+                                skip_deserialize=False,
+                            ),
+                        ),
+                    ),
+                    (
+                        'nested_empty_named_tuple',
+                        tree_metadata.ValueMetadataEntry(
+                            value_type='None',
+                            skip_deserialize=True,
+                        ),
+                    ),
+                ],
+            )
         },
     ),
 ]